Исправлена ошибка: был неверно объявлен прототип функции

+++ src/discriminant.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 #include <math.h>
 
-double discriminant(int a, int b, int c)
+long double discriminant(int a, int b, int c)
 {
 	long double d;
 	d = b * b - 4 * a * c;
